refactor(go-generate): Use qualified name for schema paths
Move getQualifiedClassName / getQualifiedInterfaceName into utils.ts and
reuse for getSchemaPath, so the local _schemas filenames now use the
same DNS-reversed PascalCase convention as the interface aliases (e.g.
ComGithubPrometheusOperator...V1Prometheus.js). The external-schema
mapping collapses to a {prefix → dir} table since the filename suffix
is now derived directly from the ref id.
